Move scaffold logic into separate module
