+++ README.md
@@ -87,8 +87,6 @@ The various `cache` folders and `config` folders should not be included.
 	        (_addGlobal_, _addOverlay_, _addCavePortrait_)
 		- `brushes`: Results dealing with instances.  
 			(_GenRotatingEnt_, _AlterFace_, _AddBrush_, _TemplateBrush_)
-	    - `cutoutTile`: Logic for the Cutout Tile item. (_CutoutTile_)
-		- `catwalks`: Logic for Catwalk items (_MakeCatwalk_)
 		- `fizzler`: Results for custom fizzler items.
 			(_CustFizzler_, _fizzlerModelPair_)
 	    - `globals`: Global flags allowing reference to stylevars, voicelines, etc.  
@@ -103,6 +101,10 @@ The various `cache` folders and `config` folders should not be included.
 		- `randomise`: Results for randomising instances.
 			(_random_, _variant_, _randomNum_, _randomVec_)
 		- `trackPlat`: Result for modifying track platforms. (_trackPlatform_)
+		
+	    - `cutoutTile`: Logic for the Cutout Tile item. (_CutoutTile_)
+		- `catwalks`: Logic for Catwalk items (_MakeCatwalk_)
+		- `scaffold`: Logic for Unstationary Scaffolds.
 	- `instanceLocs`: Translates `<ITEM_ID:0,1>` text into the associated instance paths.
 	- `vbsp`: The BEE2's VBSP hook, which modifies a map VMF before it is compiled by the original VBSP.
 	- `vbsp_launch`: Wrapper around vbsp, to get around the renaming of scripts to `'__main__'`.
